+++ ChangeLog.md
@@ -13,7 +13,7 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
   - `onEmailAlreadyExist` to specify an action if the email is already registered.
   - `confirmationEmailResentTemplate` to notify user that a confirmation email has been resent
 * Route `getConfirmationEmailResentR` to redirect user after the confirmation email has been resent
-* Functions `getError`, `setError`, `cleanError` added to export list
+* Functions `getError`, `setError`, `clearError` added to export list
 
 ### Changed